move BeamRecord to sdk/core
