+++ sdks/java/extensions/sql/src/main/java/org/apache/beam/sdk/extensions/sql/schema/BeamSqlRowCoder.java
@@ -20,6 +20,7 @@ package org.apache.beam.sdk.extensions.sql.schema;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.sql.Types;
 import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.List;
@@ -34,13 +35,12 @@ import org.apache.beam.sdk.coders.DoubleCoder;
 import org.apache.beam.sdk.coders.InstantCoder;
 import org.apache.beam.sdk.coders.ListCoder;
 import org.apache.beam.sdk.coders.StringUtf8Coder;
-import org.apache.beam.sdk.extensions.sql.impl.utils.CalciteUtils;
 
 /**
  *  A {@link Coder} encodes {@link BeamSqlRow}.
  */
 public class BeamSqlRowCoder extends CustomCoder<BeamSqlRow> {
-  private BeamSqlRowType tableSchema;
+  private BeamSqlRowType sqlRecordType;
 
   private static final ListCoder<Integer> listCoder = ListCoder.of(BigEndianIntegerCoder.of());
 
@@ -52,58 +52,59 @@ public class BeamSqlRowCoder extends CustomCoder<BeamSqlRow> {
   private static final BigDecimalCoder bigDecimalCoder = BigDecimalCoder.of();
   private static final ByteCoder byteCoder = ByteCoder.of();
 
-  public BeamSqlRowCoder(BeamSqlRowType tableSchema) {
-    this.tableSchema = tableSchema;
+  public BeamSqlRowCoder(BeamSqlRowType sqlRecordType) {
+    this.sqlRecordType = sqlRecordType;
   }
 
   @Override
-  public void encode(BeamSqlRow value, OutputStream outStream) throws CoderException, IOException {
+  public void encode(BeamSqlRow value, OutputStream outStream)
+      throws CoderException, IOException {
     listCoder.encode(value.getNullFields(), outStream);
     for (int idx = 0; idx < value.size(); ++idx) {
       if (value.getNullFields().contains(idx)) {
         continue;
       }
 
-      switch (CalciteUtils.getFieldType(value.getDataType(), idx)) {
-        case INTEGER:
+      switch (sqlRecordType.getFieldsType().get(idx)) {
+        case Types.INTEGER:
           intCoder.encode(value.getInteger(idx), outStream);
           break;
-        case SMALLINT:
+        case Types.SMALLINT:
           intCoder.encode((int) value.getShort(idx), outStream);
           break;
-        case TINYINT:
+        case Types.TINYINT:
           byteCoder.encode(value.getByte(idx), outStream);
           break;
-        case DOUBLE:
+        case Types.DOUBLE:
           doubleCoder.encode(value.getDouble(idx), outStream);
           break;
-        case FLOAT:
+        case Types.FLOAT:
           doubleCoder.encode((double) value.getFloat(idx), outStream);
           break;
-        case DECIMAL:
+        case Types.DECIMAL:
           bigDecimalCoder.encode(value.getBigDecimal(idx), outStream);
           break;
-        case BIGINT:
+        case Types.BIGINT:
           longCoder.encode(value.getLong(idx), outStream);
           break;
-        case VARCHAR:
-        case CHAR:
+        case Types.VARCHAR:
+        case Types.CHAR:
           stringCoder.encode(value.getString(idx), outStream);
           break;
-        case TIME:
+        case Types.TIME:
           longCoder.encode(value.getGregorianCalendar(idx).getTime().getTime(), outStream);
           break;
-        case DATE:
-        case TIMESTAMP:
+        case Types.DATE:
+        case Types.TIMESTAMP:
           longCoder.encode(value.getDate(idx).getTime(), outStream);
           break;
-        case BOOLEAN:
+        case Types.BOOLEAN:
           byteCoder.encode((byte) (value.getBoolean(idx) ? 1 : 0), outStream);
           break;
 
         default:
           throw new UnsupportedOperationException(
-              "Data type: " + value.getDataType().getFieldsType().get(idx) + " not supported yet!");
+              "Data type: " + sqlRecordType.getFieldsType().get(idx) + " not supported yet!");
       }
     }
 
@@ -115,55 +116,55 @@ public class BeamSqlRowCoder extends CustomCoder<BeamSqlRow> {
   public BeamSqlRow decode(InputStream inStream) throws CoderException, IOException {
     List<Integer> nullFields = listCoder.decode(inStream);
 
-    BeamSqlRow record = new BeamSqlRow(tableSchema);
+    BeamSqlRow record = new BeamSqlRow(sqlRecordType);
     record.setNullFields(nullFields);
-    for (int idx = 0; idx < tableSchema.size(); ++idx) {
+    for (int idx = 0; idx < sqlRecordType.size(); ++idx) {
       if (nullFields.contains(idx)) {
         continue;
       }
 
-      switch (CalciteUtils.getFieldType(tableSchema, idx)) {
-        case INTEGER:
+      switch (sqlRecordType.getFieldsType().get(idx)) {
+        case Types.INTEGER:
           record.addField(idx, intCoder.decode(inStream));
           break;
-        case SMALLINT:
+        case Types.SMALLINT:
           record.addField(idx, intCoder.decode(inStream).shortValue());
           break;
-        case TINYINT:
+        case Types.TINYINT:
           record.addField(idx, byteCoder.decode(inStream));
           break;
-        case DOUBLE:
+        case Types.DOUBLE:
           record.addField(idx, doubleCoder.decode(inStream));
           break;
-        case FLOAT:
+        case Types.FLOAT:
           record.addField(idx, doubleCoder.decode(inStream).floatValue());
           break;
-        case BIGINT:
+        case Types.BIGINT:
           record.addField(idx, longCoder.decode(inStream));
           break;
-        case DECIMAL:
+        case Types.DECIMAL:
           record.addField(idx, bigDecimalCoder.decode(inStream));
           break;
-        case VARCHAR:
-        case CHAR:
+        case Types.VARCHAR:
+        case Types.CHAR:
           record.addField(idx, stringCoder.decode(inStream));
           break;
-        case TIME:
+        case Types.TIME:
           GregorianCalendar calendar = new GregorianCalendar();
           calendar.setTime(new Date(longCoder.decode(inStream)));
           record.addField(idx, calendar);
           break;
-        case DATE:
-        case TIMESTAMP:
+        case Types.DATE:
+        case Types.TIMESTAMP:
           record.addField(idx, new Date(longCoder.decode(inStream)));
           break;
-        case BOOLEAN:
+        case Types.BOOLEAN:
           record.addField(idx, byteCoder.decode(inStream) == 1);
           break;
 
         default:
           throw new UnsupportedOperationException("Data type: "
-              + CalciteUtils.toCalciteType(tableSchema.getFieldsType().get(idx))
+              + sqlRecordType.getFieldsType().get(idx)
               + " not supported yet!");
       }
     }
@@ -174,8 +175,8 @@ public class BeamSqlRowCoder extends CustomCoder<BeamSqlRow> {
     return record;
   }
 
-  public BeamSqlRowType getTableSchema() {
-    return tableSchema;
+  public BeamSqlRowType getSqlRecordType() {
+    return sqlRecordType;
   }
 
   @Override
